feat: complete chapter7 senior mission
