+++ src/main/java/umc/apiPayload/exception/ExceptionAdvice.java
@@ -4,6 +4,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.env.Environment;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
@@ -12,30 +13,29 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.client.HttpServerErrorException;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 import umc.apiPayload.ApiResponse;
 import umc.apiPayload.code.ErrorReasonDTO;
 import umc.apiPayload.code.status.ErrorStatus;
-import umc.webhook.DiscordClient;
+import umc.webhook.client.DiscordClient;
 import umc.webhook.DiscordMessage;
+import umc.webhook.service.DiscordAlarmService;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.time.LocalDateTime;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 
 @Slf4j
 @RestControllerAdvice(annotations = {RestController.class})
 @RequiredArgsConstructor
 public class ExceptionAdvice extends ResponseEntityExceptionHandler {
 
-    // FeignClient 빈을 주입
-    private final DiscordClient discordClient;
+    private final DiscordAlarmService alarmService;
+    private final Environment environment;
 
     @ExceptionHandler
     public ResponseEntity<Object> validation(ConstraintViolationException e, WebRequest request) {
@@ -67,10 +67,23 @@ public class ExceptionAdvice extends ResponseEntityExceptionHandler {
     public ResponseEntity<Object> exception(Exception e, WebRequest request) {
         e.printStackTrace();
 
-        sendDiscordAlarm(e, request);
+        if (!Arrays.asList(environment.getActiveProfiles()).contains("local")) {
+            //sendDiscordAlarm(e, request);
+            alarmService.notify(e, request);
+        }
+
 
         return handleExceptionInternalFalse(e, ErrorStatus._INTERNAL_SERVER_ERROR, HttpHeaders.EMPTY, ErrorStatus._INTERNAL_SERVER_ERROR.getHttpStatus(),request, e.getMessage());
     }
+
+//    @ExceptionHandler(HttpServerErrorException.InternalServerError.class)
+//    public ResponseEntity<Object> internalServerException(HttpServerErrorException.InternalServerError e, WebRequest request) {
+//        e.printStackTrace();
+//
+//        alarmService.notify(e, request);
+//
+//        return handleExceptionInternalFalse(e, ErrorStatus._INTERNAL_SERVER_ERROR, HttpHeaders.EMPTY, ErrorStatus._INTERNAL_SERVER_ERROR.getHttpStatus(),request, e.getMessage());
+//    }
 
     @ExceptionHandler(value = GeneralException.class)
     public ResponseEntity onThrowException(GeneralException generalException, HttpServletRequest request) {
@@ -128,51 +141,5 @@ public class ExceptionAdvice extends ResponseEntityExceptionHandler {
                 errorCommonStatus.getHttpStatus(),
                 request
         );
-    }
-
-    private void sendDiscordAlarm(Exception e, WebRequest request) {
-        discordClient.sendAlarm(createMessage(e, request));
-    }
-
-    private DiscordMessage createMessage(Exception e, WebRequest request) {
-        return DiscordMessage.builder()
-                .content("# 🚨 에러 발생 비이이이이사아아아앙")
-                .embeds(
-                        List.of(
-                                DiscordMessage.Embed.builder()
-                                        .title("ℹ️ 에러 정보")
-                                        .description(
-                                                "### 🕖 발생 시간\n"
-                                                        + LocalDateTime.now()
-                                                        + "\n"
-                                                        + "### 🔗 요청 URL\n"
-                                                        + createRequestFullPath(request)
-                                                        + "\n"
-                                                        + "### 📄 Stack Trace\n"
-                                                        + "```\n"
-                                                        + getStackTrace(e).substring(0, 1000)
-                                                        + "\n```")
-                                        .build()
-                        )
-                )
-                .build();
-    }
-
-    private String createRequestFullPath(WebRequest webRequest) {
-        HttpServletRequest request = ((ServletWebRequest) webRequest).getRequest();
-        String fullPath = request.getMethod() + " " + request.getRequestURL();
-
-        String queryString = request.getQueryString();
-        if (queryString != null) {
-            fullPath += "?" + queryString;
-        }
-
-        return fullPath;
-    }
-
-    private String getStackTrace(Exception e) {
-        StringWriter stringWriter = new StringWriter();
-        e.printStackTrace(new PrintWriter(stringWriter));
-        return stringWriter.toString();
     }
 }
