+++ src/main/java/umc/webhook/service/DiscordAlarmService.java
@@ -1,0 +1,8 @@
+package umc.webhook.service;
+
+import org.springframework.web.context.request.WebRequest;
+
+public interface DiscordAlarmService {
+
+    void notify(Exception e, WebRequest request);
+}
